tools/web_scraper_agent.py
from typing import Dict, List, Optional, Tuple
from enum import Enum

import requests
from PIL import Image
from selenium import webdriver
from selenium.webdriver.chrome.options import Options
from selenium.webdriver.common.by import By
from selenium.webdriver.support.ui import WebDriverWait
from selenium.webdriver.support import expected_conditions as EC

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler("social_growth.log"),
        logging.StreamHandler(sys.stdout)
    ]
)

class Platform(Enum):
    INSTAGRAM = "instagram"
    TIKTOK = "tiktok"
    TWITTER = "twitter"

@dataclass
class AccountConfig:
    platform: Platform
    target_countries: List[str]
    num_accounts: int
    profile_pictures: List[str]

class SocialMediaGrowthTool:
    def __init__(self):
        self.platform_configs = {}
        self.running = False
        self.results_queue = queue.Queue()
        self.proxy_manager = ProxyManager()
        self.name_generators = {
            Platform.INSTAGRAM: self._generate_instagram_name,
            Platform.TIKTOK: self._generate_tiktok_name,
            Platform.TWITTER: self._generate_twitter_name
        }
        
    def configure_platform(self, platform: Platform, config: AccountConfig):
        """Configure platform parameters"""
        self.platform_configs[platform] = config
        logging.info(f"Configured {platform.value} with {config.num_accounts} accounts")
    
    def _generate_instagram_name(self) -> str:
        """Generate Instagram-compatible name"""
        first_names = ["Adams", "Chukwu", "Efe", "Fatou", "Kofi", "Lerato"]
        last_names = ["Okafor", "Mensah", "Sowah", "Mabena", "Nkosi", "Khumalo"]
        return f"{random.choice(first_names)}_{random.choice(last_names)}"
    
    def _generate_tiktok_name(self) -> str:
        """Generate TikTok-compatible name"""
        first_names = ["Adams", "Chukwu", "Efe", "Fatou", "Kofi", "Lerato"]
        last_names = ["Okafor", "Mensah", "Sowah", "Mabena", "Nkosi", "Khumalo"]
        return f"{random.choice(first_names)}{random.choice(last_names)}"
    
    def _generate_twitter_name(self) -> str:
        """Generate Twitter-compatible name"""
        first_names = ["Adams", "Chukwu", "Efe", "Fatou", "Kofi", "Lerato"]
        last_names = ["Okafor", "Mensah", "Sowah", "Mabena", "Nkosi", "Khumalo"]
        return f"{random.choice(first_names)}{random.choice(last_names)}"
    
    def _get_random_country(self) -> str:
        """Get random target country"""
        all_countries = ["Nigeria", "India", "USA", "UK", "Brazil", "South Africa"]
        return random.choice(all_countries)
    
    def _get_random_proxy(self) -> str:
        """Get random available proxy"""
        proxies = self.proxy_manager.get_proxies()
        if not proxies:
            raise Exception("No proxies available")
        return random.choice(proxies)
    
    def _setup_selenium_driver(self, proxy: str) -> webdriver.Chrome:
        """Setup Chrome driver with proxy"""
        options = Options()
        options.add_argument("--disable-blink-features=AutomationControlled")
        options.add_experimental_option("excludeSwitches", ["enable-automation"])
        options.add_experimental_option("useAutomationExtension", False)
        options.add_argument(f"--proxy-server={proxy}")
        
        driver = webdriver.Chrome(options=options)
        driver.execute_script("Object.defineProperty(navigator, 'webdriver', {get: () => undefined})")
        return driver
    
    def _create_account(self, platform: Platform, proxy: str) -> Optional[Dict]:
        """Create account on specified platform"""
        try:
            driver = self._setup_selenium_driver(proxy)
            email = f"user_{int(time.time())}@example.com"
            username = f"user_{random.randint(10000, 99999)}"
            password = f"P@ssw0rd{random.randint(1000, 9999)}"
            
            # Platform-specific account creation
            if platform == Platform.INSTAGRAM:
                # Instagram-specific logic
                pass
            elif platform == Platform.TIKTOK:
                # TikTok-specific logic
                pass
            elif platform == Platform.TWITTER:
                # Twitter-specific logic
                pass
            
            driver.quit()
            return {
                "platform": platform.value,
                "email": email,
                "username": username,
                "password": password,
                "proxy": proxy
            }
        except Exception as e:
            logging.error(f"Error creating account: {str(e)}")
            return None
    
    def _grow_account(self, account: Dict) -> bool:
        """Grow account with engagement"""
        try:
            proxy = account["proxy"]
            driver = self._setup_selenium_driver(proxy)
            
            # Platform-specific growth logic
            if account["platform"] == "instagram":
                # Instagram growth logic
                pass
            elif account["platform"] == "tiktok":
                # TikTok growth logic
                pass
            elif account["platform"] == "twitter":
                # Twitter growth logic
                pass
            
            driver.quit()
            return True
        except Exception as e:
            logging.error(f"Error growing account: {str(e)}")
            return False
    
    def _process_account(self, platform: Platform, config: AccountConfig) -> None:
        """Process single account for growth"""
        for _ in range(config.num_accounts):
            if not self.running:
                break
                
            try:
                proxy = self._get_random_proxy()
                account = self._create_account(platform, proxy)
                
                if account and self._grow_account(account):
                    self.results_queue.put({
                        "status": "success",
                        "account": account
                    })
                else:
                    self.results_queue.put({
                        "status": "failed",
                        "account": account
                    })
                
                # Random delay to avoid detection
                time.sleep(random.uniform(180, 420))
            except Exception as e:
                logging.error(f"Error processing account: {str(e)}")
                self.results_queue.put({
                    "status": "error",
                    "error": str(e)
                })
    
    def start_growth(self) -> None:
        """Start growth process"""
        self.running = True
        self.results_queue = queue.Queue()
        
        threads = []
        for platform, config in self.platform_configs.items():
            if not config.target_countries:
                continue
                
            thread = threading.Thread(
                target=self._process_account,
                args=(platform, config),
                daemon=True
            )
            threads.append(thread)
            thread.start()
        
        # Monitor results
        while self.running:
            try:
                result = self.results_queue.get(timeout=1)
                logging.info(f"Account result: {result}")
            except queue.Empty:
                continue
    
    def stop_growth(self) -> None:
        """Stop growth process"""
        self.running = False
        logging.info("Growth stopped")

class ProxyManager:
    def __init__(self):
        self.proxies = []
        self.load_proxies()
    
    def load_proxies(self) -> None:
        """Load proxies from configuration file"""
        try:
            with open("proxies.json", "r") as f:
                data = json.load(f)
                self.proxies = data.get("proxies", [])
        except Exception as e:
            logging.error(f"Error loading proxies: {str(e)}")
    
    def get_proxies(self) -> List[str]:
        """Get list of available proxies"""
        return self.proxies.copy()

if __name__ == "__main__":
    tool = SocialMediaGrowthTool()
    
    # Configure platforms
    instagram_config = AccountConfig(
        platform=Platform.INSTAGRAM,
        target_countries=["Nigeria", "India"],
        num_accounts=100,
        profile_pictures=[]
    )
    tool.configure_platform(Platform.INSTAGRAM, instagram_config)
    
    # Start growth
    try:
        tool.start_growth()
    except KeyboardInterrupt:
        tool.stop_growth()